refactor: move consumer logic to dedicated package
- Move consumer implementation from pkg/deps to cmd/consumer
- Update main.go to use new consumer package location
- Remove consumer.go from pkg/deps package

This change improves code organization by moving consumer logic to a
dedicated package in cmd/, separating it from infrastructure dependencies.
